Update PySight.py
Commented out line 344. Opened a bug with CIRCL on the default action of publish. According to documentation, the default action is not to send email. From https://pymisp.readthedocs.io/en/latest/modules.html
publish(event, alert=False)[source]
Publish the event with one single HTTP POST. The default is to not send a mail as it is assumed this method is called on update.